move padded batch serialization tests to kernel tests
