+++ tensorflow/python/data/kernel_tests/BUILD
@@ -581,6 +581,7 @@ tf_py_test(
     size = "small",
     srcs = ["padded_batch_test.py"],
     deps = [
+        ":checkpoint_test_base",
         ":test_base",
         "//tensorflow/python:array_ops",
         "//tensorflow/python:client_testlib",
